+++ app/admin/pledge/detail/controller.js
@@ -12,8 +12,10 @@ export default Ember.Controller.extend({
 
     deletePledge() {
       const no = this.model;
-      no.destroyRecord();
-      this.transitionToRoute('admin.pledge');
+      if (confirm('Are you for real?')) {
+        no.destroyRecord();
+        this.transitionToRoute('admin.pledge');
+      }
     }
 
   },
